busybox: fdisk: fix non-512 sector size disks (GPT)
 ref.: http://lists.busybox.net/pipermail/busybox/2011-March/075325.html

 The latest busybox fdisk utility with GPT enabled displays the wrong
 size for the disk and partitions. This appears to be because
 smart_ulltoa5() is expecting bytes, where gpt_list_table() is sending
 sectors.  The attached patch fixes it for me.

+++ release/src/router/busybox/util-linux/fdisk_gpt.c
@@ -93,7 +93,8 @@ gpt_list_table(int xtra UNUSED_PARAM)
 	int i;
 	char numstr6[6];
 
-	smart_ulltoa5(total_number_of_sectors * sector_size, numstr6, " KMGTPEZY")[0] = '\0';
+	smart_ulltoa5((unsigned long long)total_number_of_sectors * sector_size,
+		numstr6, " KMGTPEZY")[0] = '\0';
 	printf("Disk %s: %llu sectors, %s\n", disk_device,
 		(unsigned long long)total_number_of_sectors,
 		numstr6);
